Initial version of app
This completes the initial version of the app which has three linked views and a data structure that conforms to Codable. The data structure is a class that also conforms to ObservableObject which means that the encoding and decoding requires custom implementation. The final element in this version of the app is a network connection to an echo service that allows us to POST a request and receive the response which allows us to test our encoding / decoding in a more realistic environment.